Refactor (#6)
* Move routes to individual files

* Enable CORS for API endpoints

* Add app management to admin panel

* Add ADMIN_STEAM_ID configuration option
