+++ app/models/app.js
@@ -1,11 +1,15 @@
 var mongoose = require('mongoose');
+var Schema = mongoose.Schema;
+var autoIncrement = require('mongoose-auto-increment');
 
-var appSchema = mongoose.Schema({
+var appSchema = Schema({
   publicId: Number,
   apiKey: String,
   name: String,
-  developer: String,
+  developer: { type: Schema.Types.ObjectId, ref: 'User' },
   type: Number
 });
+
+appSchema.plugin(autoIncrement.plugin, { model: 'App', field: 'publicId', startAt: 1 });
 
 module.exports = mongoose.model('App', appSchema);
