+++ app/routes/api/v2/login.js
@@ -1,0 +1,4 @@
+module.exports = (req, res, next) => {
+  req.session.redirect = '/api/v2/' + req.params['id'] + '/result';
+  next();
+};
