+++ app/routes/developer.js
@@ -1,0 +1,5 @@
+module.exports = (req, res) => {
+  res.render('dev.ejs', {
+    user: req.user
+  });
+};
