+++ app/routes/empty.js
@@ -1,0 +1,3 @@
+module.exports = (req, res) => {
+  
+};
