+++ app/routes/help.js
@@ -1,0 +1,5 @@
+module.exports = (req, res) => {
+  res.render('help.ejs', {
+    user: req.user
+  });
+};
