+++ app/routes/logout.js
@@ -1,0 +1,4 @@
+module.exports = (req, res) => {
+  req.logout();
+  res.redirect('/');
+};
