+++ app/routes/steam-callback.js
@@ -1,0 +1,4 @@
+module.exports = (req, res) => {
+  res.redirect(req.session.redirect || '/profile');
+  delete req.session.redirect;
+};
