+++ package.json
@@ -23,11 +23,13 @@
     "connect-flash": "~0.1.1",
     "cookie-parser": "~1.0.1",
     "cookie-session": "^2.0.0-alpha.2",
+    "cors": "^2.8.3",
     "ejs": "~0.8.5",
     "express": "~4.0.0",
     "jsonwebtoken": "^7.2.1",
     "method-override": "~1.0.0",
     "mongoose": "~4.4",
+    "mongoose-auto-increment": "^5.0.1",
     "morgan": "~1.0.0",
     "passport": "^0.3.2",
     "passport-steam": "^1.0.7",
